Revert "DAOS-3527 Placement: Create Placement Map Unit Tests (#1347)"
This reverts commit 236ab5e52404cccc9b7ebbabf9525fc1719a2a63.
