+++ src/placement/SConscript
@@ -15,8 +15,7 @@ def scons():
 
     # Common placement code
     common_tgts = denv.SharedObject(['pl_map.c', 'ring_map.c',
-                                     'jump_map.c', 'pl_map_common.c',
-				     'tests/place_obj_common.c'])
+                                     'jump_map.c', 'pl_map_common.c'])
 
     # generate server module
     srv = daos_build.library(denv, 'placement', common_tgts)
