+++ src/placement/tests/SConscript
@@ -9,19 +9,11 @@ def scons():
 
     denv.AppendUnique(RPATH=[Literal(r'\$$ORIGIN/../lib/daos_srv')])
 
-    ring_test_tgt = denv.SharedObject('ring_map_place_obj.c')
-    jump_test_tgt = denv.SharedObject('jump_map_place_obj.c')
-
-    ring_pl_test = daos_build.program(denv, 'ring_pl_map', ring_test_tgt + common_tgts,
+    test_tgt = denv.SharedObject('place_obj.c')
+    pl_test = daos_build.program(denv, 'pl_map', test_tgt + common_tgts,
                                  LIBS=['daos', 'daos_common', 'gurt', 'cart',
                                        'placement', 'uuid', 'pthread', 'isal'])
-
-    jump_pl_test = daos_build.program(denv, 'jump_pl_map', jump_test_tgt + common_tgts,
-                                 LIBS=['daos', 'daos_common', 'gurt', 'cart',
-                                       'placement', 'uuid', 'pthread', 'isal'])
-
-    denv.Install('$PREFIX/bin/', ring_pl_test)
-    denv.Install('$PREFIX/bin/', jump_pl_test)
+    denv.Install('$PREFIX/bin/', pl_test)
 
 if __name__ == "SCons.Script":
     scons()
