+++ src/tests/ftest/unittest/unittest.py
@@ -80,21 +80,13 @@ class UnitTestWithoutServers(Test):
         """
         unittest_runner(self, "vea_ut")
 
-    def test_ring_pl_map(self):
+    def test_pl_map(self):
         """
-        Test Description: Test ring_pl_map unittest.
-        Use Case: This tests the ring placement map
-        :avocado: tags=all,unittest,tiny,regression,vm,ring_pl_map
+        Test Description: Test pl_map unittest.
+        Use Case: This tests placement map
+        :avocado: tags=all,unittest,tiny,regression,vm,pl_map
         """
-        unittest_runner(self, "ring_pl_map")
-
-    def test_jump_pl_map(self):
-        """
-        Test Description: Test jump_pl_map unittest.
-        Use Case: This tests the jump placement map
-        :avocado: tags=all,unittest,tiny,regression,vm,jump_pl_map
-        """
-        unittest_runner(self, "jump_pl_map")
+        unittest_runner(self, "pl_map")
 
     @skipForTicket("DAOS-1763")
     def test_eq_tests(self):
